ray: introduce a rudimentary ray tracer
My first ray tracer, it only has spheres, planes, and point light sources.

No texture mapping, no soft shadows, no global illumination.

This is all very basic right now, the camera movement is simple and boring, but
sufficient for further development and optimization.

I made some effort to support multiple CPUs, it should detect the number of
CPUs in the system and use enough pthreads to keep them busy.

Jacco Bikker's tutorial on flipcode was the original impetus to do this, and
definitely served as a guide early on.