sparkler: add per-bsp (last) lookup cache
This is a trivial optimization which makes a substantial difference.

When we're doing things like explosions, thousands of occupants get
added to the bsp tree at the exact same position.  Without the lookup
cache we end up traversing the octree down to its maximum depthrepeatedly,
because the bv containing the explosion is of course full and maximally
deep.

Now explosions don't need to spend a bunch of time in the octree just to
keep locating the same full bv the explosion occurs in.