fb: add frame_{width,height} to fb_fragment_t
Modules need to know the overall dimensions of the frame the fragment
they're rendering is part of.  Previously it wasn't really necessary
since the fragments supplied to the modules had always been the full
page, but that's changing.

This commit also changes the julia module to use the frame dimensions,
others will need updating as well.