sparkler: respect fragment->frame_{width,height}
should_draw_expire_if_oob() assumed the fragment represented the entire
frame.  Instead, return 0 if the coordinates are outside the fragment,
but only reset longevity if outside of the frame.

If sparkler goes threaded in the drawing, this would result in threads
simply skipping particles outside the fragment.  The longevity reset
occurring in all threads appears suspicious but should be benign since
they all write the same thing - 0.