*: use fragment generator
Rather than laying out all fragments in a frame up-front in
ray_module_t.prepare_frame(), return a fragment generator
(rototiller_fragmenter_t) which produces the numbered fragment
as needed.

This removes complexity from the serially-executed
prepare_frame() and allows the individual fragments to be
computed in parallel by the different threads.  It also
eliminates the need for a fragments array in the
rototiller_frame_t, indeed rototiller_frame_t is eliminated
altogether.