rototiller: introduce module.finish_frame()
Add a hook for post-render serialized frame completion,
some of the renderers may have state to cleanup after rendering
a frame.

A future commit may change add a return value to control flow for
features like multi-pass rendering within a given module.

The raytracer for example may want to add concurrently executed
post filters, and having a non-void return from finish_frame()
would be a tidy way to tell rototiller "go back to prepare->render
for this context" as many times as necessary, keeping the pass state
in the context.

For now its return is void however, as I just need a cleanup hook
as the raytracer becomes more stateful per frame with a BIH spatial
index in the works.