sdl_fb: implement rudimentary sdl fb backend
This uses a simple fixed 640x480 windowed mode (for now).

The SDL2 Renderer & Texture API is used for vsync-synchronized presents.
There's probably excessive copying going on because the rototiller fb
code manages pages and flips but SDL2 doesn't really expose low-level
control of such things.

This backend is quite useful for development purposes, allowing quick
iteration in a windowed environment.

Note this is just the backend implementation, it's dormant code but
trivially activated.