rototiller,fb: swap dispatch with page flipping
For the sake of sdl_fb, move page flipping into the main thread
and run module render dispatch from another thread instead.

This eliminates the fb flipper thread, moving its functionality
into fb_flip() which synchronously consumes and performs a single
flip from the same queue as before - the function is verbatim
the loop body of the flipper thread.

Now main() calls fb_flip() in a loop where it previously dispatched
pages for rendering.

Rendering dispatch is now performed in a created thread.

See the comment in fb.c for more explanation of this shuffle.