sdl_fb: add fullscreen= and size= settings
fullscreen takes either "on" or "off"

size expects WxH arguments, defaults still to 640x480

size is optional when fullscreen=on.

Through the setup machinery, when fullscreen has been selected it will
not ask for a size - a "fullscreen desktop" mode is presumed.

However, thorugh the explicit commandline flags, a mixed mode can be
achieved by specifying both "fullscreen=on,size=WxH".  This instructs
SDL to attempt a video mode switch to the specified size if needed.

I've found it to be pretty unreliable on my Xorg/linux system, unless I
choose the same video mode as my desktop is already in.  Then I get what
looks like rendering into the root window or something, it's weird.

Hence there's no effort made to expose that in the interactive setup, but
it's technically possible and some effort was made to wire it up.