modules/submit: add cellular automata game module
This module displays realtime battle for domination simulated
as 2D cellular automata.

This is just a test of the backend piece for a work-in-progress
multiplayer game idea.  The visuals were kind of interesting to
watch so I figured may as well merge it as a module to share.

Enjoy!

PS: the results can vary a lot by tweaking the defines in submit.c