modules/submit: add bilinearly-interpolated variant
This substantially reworks the cell sampling in submit.

As a result, it's now threaded in the rendering phase which now
resembles a texture mapper sans transformations.

This produces a full-screen rendering rather than a potentially
smaller one when the resolution wasn't cleanly divisable by the grid
size.

A new module, named submit-softly has also been added to expose the
bilinearly interpolated variant.  The transition between cells is also
employing a smoothstep so it's not actually linear.

The original non-interpolated version is retained as well, at the same
submit module name.

Some minor cleanups happened as well, nothing worth mentioning, except
perhaps that the cells are now a uint8_t which is fine unless someone
tries to redefine NUM_PLAYERS > 255.