rototiller: setup_from_args() has defaults already
I think passing this separately is vestigial from before there was an
args struct encapsulating everything.

In the future there might be some defaults discretion supported to say
use defaults for module settings but not video, or vice versa.  So get
rid of this pointless parameter in prep for that, just use the args
struct.