flui2d: add some rudimentary settings
Viscosity and diffusion are supported, it'd be neat to add a
configurable size (the ROOT define) for the flow field in the
future.

I didn't go crazy here, it's just a list of orders of magnitude you
choose from for each.  It'd probably be more interesting to change
this into a single knob with descriptive names like "smoke" "goop"
"water" mapping to a LUT.