rtv: implement "Rototiller TV" renderer
This is sort of a meta renderer, as it simply renders other
modules in its prepare_frame() stage.  They're still threaded
as the newly public rototiller_module_render() utilizes the
threading machinery, it just needs to be called from the serial
phase @ prepare_frame().

I'm pretty sure this module will leak memory every time it changes
modules, since the existing cleanup paths for the modules hasn't
needed to be thorough in the least.  So that's something to fix
in a later commit, go through all the modules and make sure their
destroy_context() entrypoints actually cleans everything up.

See the source for some rtv-specific TODOs.