snow: add a simple tv snow / white noise module
I wanted to add some noise to the rtv module and figured why not
just add a snow module and make rtv pass through it briefly when
switching modules.

It's not interesting by itself, but as more composite/meta modules
like rtv get made it might be handy beyond rtv.