settings: add setting_desc_check() helper
Rudimentary setting value checking against the description.

For now it just enforces the multiple-choice stuff, I'm undecided
on the regex support for now.  It'd just be nice to throw some
more informative errors when cli arguments are incorrect for things
like fullscreen=yes when it only knows fullscreen=on/off.