rototiller: make rtv the default module
Also sort the modules alphabetically.

Now that the major memory leaks are addressed (sparkler), make
the rtv module the default since it gives the user an automated
tour of all the modules.  Explicit module use is more aimed at
tinkerers playing with a specific module's code either creating
their own or modifying an existing one, but isn't really desirable
as the default flow.