libs/din: add a perlin noise implementation
This is a 3D noise field addressed as a unit cube.

The caller supplies the resolution of the noise field in three
dimensions.

I've just pulled in my v3f.h here, but it probably makes sense to
later on move vector headers into libs/ and share them.  Later.

It's called din as in noise, because it's shorter than perlin and
noise.