swab: add a perlin noise visualization
This maps a different Z-slice through the noise field to each color
channel.  The slices are moved up and down through the field over
time, and the size of the area each color samples is tweaked a bit
to make them less coherent with the noise field cells.

It could be improved, but I think the output is already neat enough
to be worth sharing.