settings: add setting_desc_t.random() method
To facilitate random setting of these flexible string-oriented
settings, support a random helper supplied with the description.

This helper would return a valid random string to be used with the
respective setting being described.

Immediate use case is the rtv module, which also gets fixed up to
use it in this commit.