fb: add pitch to fb_fragment_t
The put_pixel helpers really needed reworking to properly handle
subframe fragments modules like montage will utilize.  I had the
stride present as it's convenient for a number of modules that
maintain a buf pointer as they progress down a row, but the pitch
is more applicable to put_pixel for scaling the y coordinate.

Now there's both pitch and stride so everyone's happy with what's
most convenient for their needs.