fb: add number to fb_fragment_t
This introduces a stricter coupling and requirement for modules
supplying a fragmenter in their prepare_frame() to only receive
fragments produced by *their* fragmenter at their render_fragment().

When modules don't explicitly perform any fragmenting they can't
really make much use of this number as it will reflect an arbitrary
fragmenting pass's perspective.

But when modules do perform their own frame fragmenting, they can
assume any fragment supplied to their render function will have been
generated by it.  This needs to be enforced a bit in the code.

The current use case is montage using a fragmenter for tiling the
montage in a threaded render.  The fragment numbers map to the
modules to be rendered in the tiles.  As long as modules can assume
their fragmenter will always be what produces their fragments, this
is perfectly fine.