montage: rework module skipping
The old approach was just to get things working, it's preferable
to not have empty tiles on-screen where modules were skipped and
have all tiles be smaller to accomodate vacancies.

Now the modules list gets pruned @ context create, so the skipping
only happens once and everywhere else is looking at a modules list
and count of only the keepers.

I also added stars to the skipped modules, for now, since both stars
and pixbounce malfunction when the fragment size changes.