snow: per-cpu rand_r() seed state
Snow was already threaded, but used a global seed with rand_r() meaning
the CPUs were hammering on the same address.  There wasn't any locking or
atomics, as it isn't terribly critical when generating white noise if the
seed access is racy.  But the writes still caused cache lines to ping-pong.

This commit gives a ~15.5% speedup in my measurements on an i7-2640M.

Note without the padded union, using just an array of ints, zero gain
is realized.  I used a padding of 256 just to have some headroom, x86
is 64 but other CPUs vary, POWER9 is 128 for example.