din: don't include v3f.h in din.h
This requires a forward declaration of v3f_t and changing din()
to take a v3f_t *.

The swab module needed updating to supply a pointer type and a
v3f_t definition.

This is being done so din.h users can have their own v3f
implementations.  I might consolidate all the duplicated vector
code scattered throughout the libs and modules, but for now I'm
carrying on with the original intention of having modules be
largely self-contained.  Though the introduction of libs like
ray and din has certainly violated that a bit already.