montage: add (threaded) to description
While the montage doesn't deeply thread the per-tile/module rendering,
the per-frame rendering is threaded with a work unit granularity of
every module's tile.

Meaning every module renders its tile in a single thread, but the
tiles are all rendered in parallel.

For the most part this works, and will only work better as more modules
are added to rototiller increasing the granularity.  In the mean time
it's a bit coarse and some modules can be a lot more costly to render
than others, and there can be a shortage of modules to schedule on idle
CPUs.

It would be an interesting task to try make each module's tile get
subfragmented elastically.  I didn't make any attempt to do that, but
it might even be worthwhile on hidpi screens where even those small
tiles may have a whole lot of pixels, especially on manycore CPUs.