libs/ray: decouple film and frame dimensions
The existing code conflated the rendered frame dimensions with
what's essentially the virtual camera's film dimensions.  That
resulted in a viewing frustum depending on the rendered frame
dimensions.  Smaller frames (like in the montage module) would
show a smaller viewport into the same scene.

Now the view into the scene always shows the same viewport in
terms of the frustum dimensions for a given combination of
focal_length and film_{width,height}.

The rendered frame is essentially a sampling of the 2D plane
(the virtual film) intersecting the frustum.

Nothing is done to try enforce a specific aspect ratio or any
such magic.  The caller is expected to manage this for now, or
just ignore it and let the output be stretched when the aspect
ratio of the output doesn't match the virtual film's aspect
ratio.

In the future it might be interesting to support letter boxing or
such things for preserving the film's aspect ratio.

For now the ray module just lets things be stretched, with
hard-coded film dimensions of something approximately consistent
with the past viewport.

The ray module could make some effort to fit the hard-coded film
dimensions to the runtime aspect ratio for the frame to be
rendered, tweaking things as needed but generally preserving the
general hard-coded dimensions.  Allowing the frustum to be
minimally adjusted to fit the circumstances... that might also be
worth shoving into libray.  Something of a automatic fitting
mode for the camera.