libs/sig: square wave; sig_ops_sqr / sig_new_sqr()
This builds minimally upon the existing sine wave code,
just use the sign to drive the signal high or low.

Wikipedia shows a third state for 0 values, but that's for a -1..+1
signal.  I'm producing all 0-1 signals as it's more convenient for
this application, but it seems like it would be awkward to return
.5f for the 0 case.  So 0 is being treated as just another positive
value; high.