module/sparkler: implement some BSP drawing settings
This commit adds a few settings for visualizing the octree BSP:

show_bsp_leafs (on/off):
	Draw wireframe cubes around octree leaf nodes

show_bsp_leafs_min_depth (0,4,6,8,10):
	Set minimum octree depth for leaf nodes displayed

show_bsp_matches (on/off):
	Draw lines connecting BSP search matches

show_bsp_matches_affected_only (on/off):
	Limit drawn BSP search matches to only matches actually
	affected by the simulation

The code implementing this stuff is a bit crufty, fb_fragment_t
had to be pulled down to the sim ops for example and whether that
actually results in drawing occurring during the sim phase
depends on the config used and how the particle implementations
react to the config... it's just gross.  This matters because the
caller used to know only the draw phase touched fb_fragment_t,
and because of that the fragment was zeroed after sim and before
draw in parallel.  But now the caller needs to know if the config
would make sim do some drawing, and do the fragment zeroing
before sim instead, and skip the zero before draw to not lose
what sim drew.  It's icky, but I'll leave it for now, at least
it's isolated to the sparkler.