modules/plasma: normalize plasma dimensions
This is a quick and dirty jab at normalizing the plasma size to
be independent of the frame size.

I kind of hate this module as-is, it's tempting to discard all
the fixed point stuff and just redo it using floats.. the plasma
itself isn't that attractive as-is either.

But I have other things to work on currently, just wanted to make
it so the plasma doesn't look like a solid color in the montage
tile.