modules/swarm: implement a particles swarm module
Just a fun little swarm based loosely on 80s-era boids

It would be interesting to make stuff like the # of particles
and the weights runtime configurable, or exposed as knobs.

Using a Z-buffer for occlusions and perhaps shading by depth might
make a significant improvement on the visual quality.  It might also
be interesting to draw the particles as lines connecting their current
position with their previous, instead as pixels.  Or fat pixels like
stars...