compose,montage,rtv: drop author and license fields
These modules are meta modules, and the only place this
information is presented currently is in the rtv module captions
overlaying the visual output of unrelated modules.

So it's rather misleading to put the meta module's author and
license on-screen when what's being shown is arguably just a tiny
fraction of the meta module's contribution.

Rather than bother with constructing license and author lists at
runtime from the modules incorporated by these meta modules,
let's instead adopt a policy of meta modules omit any declaration
of license or authorship outside of the source.  This is a simple
solution for now, it can be revisited later if necessary.

Changing the .author member of rototiller_module_t to an
.authors() function pointer wouldn't be difficult.  But it does
open up something of a can of worms when considering recursive
dependencies and needing to construct unique authors and licenses
lists from things like nested meta modules.  Obviously there
can't be infinite recursion as that would manifest in the
rendering path as well, but what I'm more concerned about is
properly handling potentialy quite long lists.  It's already
annoying when rtv has to deal with a long settings string, which
I believe currently is just truncated.  The same would have to be
done with long authors/licenses I guess.

In any case, I think it's probably fine to just leave authorship
and license ambiguous when a meta module is shown in rtv.  It's
certainly preferable to vcaputo@pengaru.com getting credit for
everything shown in the three meta modules currently implemented,
or more specifically, the two shown in rtv; compose and montage.

Note this required making rtv tolerante of NULL .license and
.author rototiller_module_t members.