fb: make inactive pages list doubly-linked
This introduces a previous pointer to _fb_page_t, but it's
only used on the inactive list.

This is a preparatory commit to facilitate adding inactive pages
at the head and consuming inactive pages off the tail, without
having to always iterate the list for the new tail.

Rebuilding pages on resize needs to happen somewhere, and the
inactive list seems like the right place to do it.  And to do it
with a simple per-fb pages-needing-rebuild counter requires
turning the inactive list into a FIFO queue.