fb: introduce fb_rebuild()
When an fb has been resized, all the pages can be forced to go
through a backend free->alloc cycle by calling this function.

This process gets queued and realized as the pages pass through
the inactive state.

This does mean that whatever pages are already queued as ready
still get displayed, so it's not exactly without
racing/flickering at the fringes on window growth.

It's possible to do something with the pages in the ready queue
to combat this, but for now I'm just leaving them be to keep it
simple.  The ready pages could be stretched or something...