modules/rtv: fix "none" snow to actually blank
This manifests in the current unconfigured rtv glimmer shows, since
the default is a the "none" module when no settings are applied.

But it turns out this isn't just a glimmer problem, "none" is advertised
in the settings as a blanking alternative to snow.  So it's actually
broken in rototiller as well.

This fixes it by detecting the nil "none" module's lack of any
prepare_frame or render_fragment methods, and open coding the blanker
with a fb_fragment_zero() inline.