args: move argument parsing/help output to libtil
This is totally opt-in for libtil callers, but is a step
towards enabling uniform cli invocations across frontends.

The help side of this is particularly janky, but since what's
appropriate there is directly related to the args parsing it
seems appropriate to bring along.  The janky part is the
implicit output formatting assumptions being made, as-is it
doesn't really lend itself well to being augmented into broader
frontend help output.  Alas, this is rototiller playground, so
let's just go easy and assume frontends will largely spit out
whatever this provides - or completely replace it if appropriate.