til_settings: provide a user_data pointer w/til_setting_t
Particularly in implementing a stateful/"retained" GUI it can be
desirable to embed something like a widget pointer in a
til_setting_t once described and shown to the user.

Management of this pointer is largely nonexistant from the
libtil perspective.  It's simply initialized to NULL when a new
setting is added, and never accessed again.  100% the caller's
responsibility.

This works fine since libtil/til_settings_t only accumulates
til_setting_t entries and never removes them except when
discarding an entire til_settings_t wholesale.