til_settings: fix empty values in til_settings_new()
Previously if you supplied an empty setting value like so:
"--module=compose,layers="

The interactive setup would get itself into an infinite loop
because the layers setting is already present, but has a NULL
value.  This wasn't a NULL value, it was a "" value.

The parser should just fallthrough to the value state from the
equal state after recording the value's start pointer.  This will
result in a "" value getting allocated and assigned to the value
before the loop breaks out on the '\0' immediately following the
'='.

There are probably other edge cases which need better handling
here.