*: wire up context-specific setup instances
This is a preparatory commit for cleaning up the existing sloppy
global-ish application of settings during the iterative _setup()
call sequences.

Due to how this has evolved from a very rudimentary thing
enjoying many assumptions about there ever only being a single
module instance being configured by the settings, there's a lot
of weirdness and inconsistency surrounding module setup WRT
changes being applied instantaneously to /all/ existing and
future context's renderings of a given module vs. requiring a new
context be created to realize changes.

This commit doesn't actually change any of that, but puts the
plumbing in place for the setup methods to allocate and
initialize a private struct encapsulating the parsed and
validated setup once the settings are complete.  This opaque
setup pointer will then be provided to the associated
create_context() method as the setup pointer.  Then the created
context can configure itself using the provided setup when
non-NULL, or simply use defaults when NULL.

A future commit will update the setup methods to allocate and
populate their respective setup structs, adding the structs as
needed, as well as updating their create_context() methods to
utilize those setups.

One consequence of these changes when fully realized will be that
every setting change will require a new context be created from
the changed settings for the change to be realized.

For settings appropriately manipulated at runtime the concept of
knobs was introduced but never finished.  That will have to be
finished in the future to enable more immediate/interactive
changing of settings-like values appropriate for interactive
manipulation