doc: first stab at module-writing documentation
Things have become sufficiently mature and featureful that
attempting to describe their usage seemed worthwhile.

There's still no cleanup of setups returned in *res_setup and
that should be both fixed and documented at some point.  The
settings in general are still rather leaky as-is, even the
example in this document leaks.  But it's relatively harmless for
now.